Restored ability to build and deploy to Android via Maven.
This involves a number of fixes:

1. Maven does the right thing with dependent resources, so we can nix the
Eclipse hackery in playn/android/Shaders. The Eclipse build is already a mess,
so we should fix it properly rather than put bandaids over one or two of the
numerous gaping wounds.

2. The native library needs to be exported as a Maven artifact. For now I've
installed the artifact into the Forplay legacy repository, so it will be
publicly accessible. When we publish PlayN to Maven Central, we can deploy the
library there as well. This also required changing the name of the library, as
the library ends up in the APK with the same name as its artifactId, which is
playn-android-nativelib. The old name (playn-android) conflicted with the name
of the playn-android.jar artifact.

3. It appears not to be possible to add the native library dependency to the
playn-android POM and have the Maven Android plugin do the right thing. The
native library dependency seems to have to be added to the end-user project
POM, so I've added to the sample android POMs and the archetype.