Nix the samples meta-pom.
Having each sample project use playn-project as its parent avoids pitfalls when
building locally. Maven will look up one directory to find a parent, so if
you're in samples/showcase/android, it will find the samples/showcase POM, but
if the samples/showcase POM depends on the samples POM, then Maven freaks out
and wants you to "mvn install" the samples POM. But that's a PITA and
pointlessly clutters your local Maven repository.

The only benefit of having one top-level POM is that we can build all the
samples in a single CI build. But I'd much rather have things be more likely to
work for random developers turning up and trying to run the samples. I can set
up multiple CI builds for the samples easily enough.

In case you were thinking we could have a top-level POM anyway, and have it
build projects that didn't declare it as a parent. Ha ha ha, you're so funny.
Yeah, that would be nice, except that Maven treats that as a violation of all
that is true and good in the world and generates annoying warnings every time
you build the individual sample projects. So we're not going down that road.