Use a custom font in the text demo. Only Java & HTML, other platforms TBD.
Currently also in the HTML backend, the first time text is rendered, it doesn't show up. But
subsequent renderings work. I don't know if custom fonts are not "ready" when first accessed and
only become ready later, or what the deal is, but the behavior is consistent across FF, Safari and
Chrome.